refactor(AlgorithmWrapper): move to the configuration subpackage
