+++ tests/unit_tests/configuration/test_configuration_manager.py
@@ -9,7 +9,7 @@ import textwrap
 from brisk.configuration.configuration_manager import ConfigurationManager
 from brisk.configuration.experiment_group import ExperimentGroup
 from brisk.data.data_manager import DataManager
-from brisk.utility.algorithm_wrapper import AlgorithmWrapper
+from brisk.configuration.algorithm_wrapper import AlgorithmWrapper
 
 class TestConfigurationManager:
     def test_initialization(self, mock_regression_project):
